done DN

+++ src/main/java/com/WebProjectManagementBE/controller/DuAnController.java
@@ -2,9 +2,12 @@ package com.WebProjectManagementBE.controller;
 
 import com.WebProjectManagementBE.DTO.DuAnDTO;
 import com.WebProjectManagementBE.model.DuAn;
+import com.WebProjectManagementBE.model.NguoiDung;
+import com.WebProjectManagementBE.model.QuanLyDuAn;
 import com.WebProjectManagementBE.service.DuAnService;
 import com.WebProjectManagementBE.service.FirebaseService;
 import com.WebProjectManagementBE.service.NguoiDungService;
+import com.WebProjectManagementBE.service.QuanLyDuAnService;
 import org.apache.coyote.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -29,6 +32,9 @@ public class DuAnController {
     @Autowired
     NguoiDungService nguoiDungService;
 
+    @Autowired
+    QuanLyDuAnService quanLyDuAnService;
+
 
     @PostMapping("")
     public ResponseEntity<?> createDuAn(@RequestBody DuAnDTO duAnDTO) {
@@ -43,13 +49,23 @@ public class DuAnController {
         }
     }
 
+    @GetMapping("/{maDuAn}")
+    public ResponseEntity<?> getDuAn(@PathVariable String maDuAn) {
+        if (duAnService.isNguoiDungInDuAn(maDuAn)) {
+            return ResponseEntity.ok(duAnService.getDuAn(maDuAn));
+        }
+        else {
+            return ResponseEntity.internalServerError().body("Không được phép truy cập");
+        }
+    }
+
     @GetMapping("")
     public ResponseEntity<?> getAllDuAnUser() {
         List<Map<String, Object>> duAns = new ArrayList<>();
-        for (DuAn da : nguoiDungService.getCurrentSessionNguoiDung().getDuAnCollection()) {
+        for (QuanLyDuAn qlda : quanLyDuAnService.getDuAnCollectionByNguoiDung(nguoiDungService.getCurrentSessionNguoiDung())) {
             Map<String, Object> data = new HashMap<>();
-            data.put("owner", duAnService.getOwnerDuAn(da));
-            data.put("duan", da);
+            data.put("owner", duAnService.getOwnerDuAn(qlda.getDuAn()));
+            data.put("duan", qlda.getDuAn());
             duAns.add(data);
         }
 
@@ -58,13 +74,44 @@ public class DuAnController {
 
     @DeleteMapping("/{maDuAn}")
     public ResponseEntity<?> deleteDuAn(@PathVariable String maDuAn) {
-        duAnService.deleteDuAn(maDuAn);
-        return ResponseEntity.ok("Xoá dự án");
+        if (duAnService.isDuAnOwnedByUser(maDuAn)) {
+            duAnService.deleteDuAn(maDuAn);
+            return ResponseEntity.ok("Xoá dự án");
+        }
+        else {
+            return ResponseEntity.internalServerError().body("Không được phép truy cập");
+        }
     }
 
     @GetMapping("/listTrangThai")
     public ResponseEntity<?> getListTrangThai() {
         return ResponseEntity.ok(DuAnService.getTrangThaiValues());
+    }
+
+
+    @PostMapping("/addNguoiDungs")
+    public ResponseEntity<?> addNguoiDungsDuAn(@RequestBody Map<String, Object> request) {
+        String maDuAn = (String) request.get("maDuAn");
+        List<String> maNguoiDungs = (List<String>) request.get("maNguoiDungs");
+        if (duAnService.isDuAnOwnedByUser(maDuAn)) {
+            DuAn duAn = duAnService.getDuAn(maDuAn);
+            for (String maNguoiDung : maNguoiDungs) {
+                NguoiDung nguoiDung = nguoiDungService.findByMaNguoiDung(maNguoiDung);
+                duAnService.addNguoiDungDuAn(nguoiDung, duAn, false);
+            }
+            duAnService.updateDuAn(duAn);
+            return ResponseEntity.ok("Thêm thành công");
+        }
+        else {
+            return ResponseEntity.internalServerError().body("Không được phép truy cập");
+
+        }
+
+    }
+
+    @GetMapping("/isDuAnOwner/{maDuAn}")
+    public ResponseEntity<?> isDuAnOwner(@PathVariable String maDuAn) {
+        return ResponseEntity.ok(duAnService.isDuAnOwnedByUser(maDuAn));
     }
 
 }
